some work on tests
