+++ code/cpm_part2.py
@@ -23,9 +23,10 @@ def execute_cpm_part2(n_test=1):
     campaign = 92 
     pixel = [883, 670]
     in_directory = "../test/intermediate/"
+    in_directory_2 = "../test/intermediate/expected/"
     out_directory = "../test/output/"
     
-    pre_matrix_file = "{:}{:}-pre_matrix_xy.dat".format(in_directory, n_test)
+    pre_matrix_file = "{:}{:}-pre_matrix_xy.dat".format(in_directory_2, n_test)
     predictor_epoch_mask_file = "{:}{:}-predictor_epoch_mask.dat".format(in_directory, n_test)
     pixel_flux_file_name = '{:}{:}-pixel_flux.dat'.format(in_directory, n_test)
     pixel_mask_file_name = '{:}{:}-mask.dat'.format(in_directory, n_test)
